requirements.txt
langchain==0.3.19
langchain_chroma==0.2.2
langchain_community==0.3.18
langchain_openai==0.3.6
python-dotenv==1.0.1
streamlit==1.42.1
